Make library object global within op.js
Makes the entire OP (and by proxy discord.js) library available globally
available to scripts loaded by `op.js`. This makes it so that plugins
and commands do not need to be part of a valid npm package and makes it
easy to install `op` for global use on a system.