Change Command.prototype.isPermitted()
Commands now check for admin rights themselves by default. This is
achieved by passing admin user IDs in an array as a second parameter to
Command.prototype.isPermitted(). This makes checking for permissions on
a command easy to overide totally for inheriting classes. The only
required behavior is that the command parser will ignore messages from the
current bot or any other bot user (which prevents weird recursive
commands).